narrow _port_in/out API uint16_t -> uint8_t (#76, -60 B SDCC resident)
Z80 IN/OUT instructions are fundamentally 8-bit on the wire (port on
A0-A7).  RC702 doesn't decode A8-A15 for I/O, so the wider uint16_t
form had no functional effect.  All RC702 port values fit in uint8_t
already.  Narrowing the runtime-port API matches the actual hardware.

hal.h:
  uint8_t _port_in (uint8_t p);                    (was uint16_t p)
  void    _port_out(uint8_t p, uint8_t v);         (was uint16_t p)
  enum : uint8_t { PORT_DMA_* = 0xF4..0xFC };      (was uint16_t)

sdcc/hal.asm: rewritten for sdcccall(1) with two u8 args in registers
(port in A, val in L) -- no stack manipulation needed for v.  Body
shrinks from ~14 B to 9 B:
  __port_in  = 4 B (LD C,A / IN A,(C) / RET)
  __port_out = 5 B (LD C,A / LD A,L / OUT (C),A / RET)
The B register is left as don't-care since RC702 doesn't decode it
during I/O cycles.

clang side: no change.  address_space(2) lowers identically for u8
and u16-narrowed-to-u8 ports; the inline IN/OUT instructions stay
the same byte sequences.

Resident size:
  SDCC pio-irq:  2050 B -> 1990 B  (-60 B / -3%)
                Phase 51 cumulative: 2756 -> 1990 = -766 B / -28%
                Gap to clang:       +408 B (+25%) -> +348 B (+21%)
  clang pio-irq: 1987 non-padding bytes (unchanged)

The -60 B SDCC saving is much larger than the ~5 B helper-shrink
estimate.  Per-call-site savings compound: each previous call site
needed `push af; inc sp` (3 B) to pass v on stack; sdcccall(1) for
two u8 args puts both in registers, so callers just `ld l, val` (2 B).
With 30+ port_in/out call sites in cpnos-rom, the per-site savings
add up to ~50-55 B.

Verified at 4-cell matrix:
  SDCC pio-irq polypascal-test  51.51 s   PASS  (was 53.25 s)
  SDCC sio     polypascal-test  60.91 s   PASS  (was 60.83 s)
  clang pio-irq polypascal-test 49.51 s   PASS  (regression check)

Closes ravn/rc700-gensmedet#76.

Rules-checked:
  feedback_check_banner_timestamp (banner shows COMPILER=sdcc per cell)
  feedback_kill_stale_servers_on_test_target (cleaned per cell)
  feedback_dual_compiler_test (both compilers verified separately)
  feedback_no_literal_addresses (port enums use named values)

+++ cpnos-rom/hal.h
@@ -27,23 +27,26 @@
  * Backend selection: _port_in / _port_out.
  *
  * Same signature in every backend:
- *     uint8_t _port_in (uint16_t port);
- *     void    _port_out(uint16_t port, uint8_t value);
+ *     uint8_t _port_in (uint8_t port);
+ *     void    _port_out(uint8_t port, uint8_t value);
  *
- * The uint16_t port type is wider than the Z80's 8-bit I/O bus but
- * matches the existing call sites (PORT_DMA_* are typed uint16_t for
- * historical reasons).  Backends truncate to 8 bits at the IN/OUT
- * instruction.
+ * Z80 IN/OUT instructions are fundamentally 8-bit (port on A0-A7).
+ * On RC702, A8-A15 are not decoded for I/O, so the wider type that
+ * was previously used (uint16_t, "for historical reasons") has no
+ * functional effect.  Narrowing to uint8_t matches the actual hardware
+ * (#76).  Backends use 8-bit port directly: clang's address_space(2)
+ * lowers to `IN A,(n)`/`OUT (n),A` for constants and `IN A,(C)`/
+ * `OUT (C),A` for runtime ports; SDCC's hal.asm uses the (C) form.
  * ================================================================ */
 
 #if defined(__clang__) && defined(__z80__)
 /* clang Z80 backend: address_space(2) lowers to IN/OUT directly. */
 #define __io __attribute__((address_space(2)))
-static inline uint8_t _port_in(uint16_t p) {
-    return *(volatile __io uint8_t *)p;
+static inline uint8_t _port_in(uint8_t p) {
+    return *(volatile __io uint8_t *)(uint16_t)p;
 }
-static inline void _port_out(uint16_t p, uint8_t v) {
-    *(volatile __io uint8_t *)p = v;
+static inline void _port_out(uint8_t p, uint8_t v) {
+    *(volatile __io uint8_t *)(uint16_t)p = v;
 }
 
 #elif defined(__SDCC) || defined(__SCCZ80)
@@ -53,8 +56,8 @@ static inline void _port_out(uint16_t p, uint8_t v) {
  * costs ~17 T-states per call (CALL + asm body + RET) versus ~12 T
  * inline.  Fine on boot init paths.  Hot-path SDCC code can still
  * use __sfr __at directly if it must shave the call. */
-extern uint8_t _port_in (uint16_t p);
-extern void    _port_out(uint16_t p, uint8_t v);
+extern uint8_t _port_in (uint8_t p);
+extern void    _port_out(uint8_t p, uint8_t v);
 
 #elif defined(__HITECH__) || defined(HI_TECH_C)
 /* HiTech C (zc / Hi-Tech Z80 C, ravn/hitech via ghcr.io/ravn/hitech).
@@ -69,8 +72,8 @@ extern void    _port_out(uint16_t p, uint8_t v);
 #else
 /* Host clang on macOS (CLion LSP, no real Z80 target).  Stubs let
  * the IDE parse the sources cleanly — no diagnostics, no codegen. */
-static inline uint8_t _port_in(uint16_t p) { (void)p; return 0; }
-static inline void    _port_out(uint16_t p, uint8_t v) { (void)p; (void)v; }
+static inline uint8_t _port_in(uint8_t p) { (void)p; return 0; }
+static inline void    _port_out(uint8_t p, uint8_t v) { (void)p; (void)v; }
 #endif
 
 /* ================================================================
@@ -107,8 +110,10 @@ enum : uint8_t {
     SIO_RR0_CTS           = 0x20
 };
 
-/* 8237 DMA controller — channel 2 = CRT display, channel 3 = CRT attr. */
-enum : uint16_t {
+/* 8237 DMA controller — channel 2 = CRT display, channel 3 = CRT attr.
+ * Narrowed from uint16_t to uint8_t in #76: all values fit in 8 bits;
+ * Z80 I/O is 8-bit on the wire; RC702 doesn't decode A8-A15 for I/O. */
+enum : uint8_t {
     PORT_DMA_CH2_ADDR = 0xF4,
     PORT_DMA_CH2_WC   = 0xF5,
     PORT_DMA_CH3_ADDR = 0xF6,
